pyproject.toml
[project]
name = "yt"
version = "0.1.0"
description = ""
authors = [
    {name = "sagarhm12",email = "sagarhmsagar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "portia-sdk-python[google] (>=0.7.2,<0.8.0)",
    "gspread (>=6.2.1,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-auth (>=2.40.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
